+++ modules/pytket-quantinuum/pytket/extensions/quantinuum/backends/quantinuum.py
@@ -146,6 +146,7 @@ class QuantinuumBackend(Backend):
         label: Optional[str] = "job",
         simulator: str = "state-vector",
         group: Optional[str] = None,
+        provider: Optional[str] = None,
         machine_debug: bool = False,
         _api_handler: QuantinuumAPI = DEFAULT_API_HANDLER,
     ):
@@ -160,6 +161,9 @@ class QuantinuumBackend(Backend):
         :param group: string identifier of a collection of jobs, can be used for usage
           tracking.
         :type group: Optional[str], optional
+        :param provider: select a provider for federated authentication. We currently
+            only support 'microsoft', which enables the microsoft Device Flow.
+        :type provider: Optional[str], optional
         :type simulator: str, optional
         :param _api_handler: Instance of API handler, defaults to DEFAULT_API_HANDLER
         :type _api_handler: QuantinuumAPI
@@ -176,6 +180,8 @@ class QuantinuumBackend(Backend):
         self.simulator_type = simulator
 
         self._api_handler = _api_handler
+
+        self._api_handler.provider = provider
 
     @classmethod
     def _available_devices(
@@ -372,6 +378,8 @@ class QuantinuumBackend(Backend):
         * `close_batch`: boolean flag to close the batch after the last circuit,
            default=True.
         * `wasm_file_handler`: a ``WasmFileHandler`` object for linked WASM module.
+        * `pytketpass`: a ``pytket.passes.BasePass`` intended to be applied
+           by the backend (beta feature, may be ignored).
 
         """
         circuits = list(circuits)
@@ -393,6 +401,7 @@ class QuantinuumBackend(Backend):
             "options": {
                 "simulator": self.simulator_type,
                 "error-model": noisy_simulation,
+                "tket": dict(),
             },
         }
         group = kwargs.get("group", self._group)
@@ -406,6 +415,11 @@ class QuantinuumBackend(Backend):
             basebody["cfl"] = cast(WasmFileHandler, wasm_fh)._wasm_file_encoded.decode(
                 "utf-8"
             )
+
+        pytket_pass = cast(Optional[BasePass], kwargs.get("pytketpass"))
+
+        if pytket_pass is not None:
+            basebody["options"]["tket"]["compilation-pass"] = pytket_pass.to_dict()
 
         handle_list = []
         batch_exec: Union[int, str]