+++ modules/pytket-ionq/docs/changelog.rst
@@ -1,6 +1,11 @@
 Changelog
 ~~~~~~~~~
 
+0.20.0 (July 2022)
+------------------
+
+* Updated pytket version requirement to 1.4.
+
 0.19.0 (June 2022)
 ------------------